refactor<decideCoinRandomly>: 메소드의 클래스 이동
변환 가능한 동전들의 경우의 수에서 랜덤하게 선택하는 메소드가 자판기보다는 동전 클래스에 있는 것이
더 적절할 것이라 생각했습니다.
