+++ src/main/java/vendingmachine/domain/Coin.java
@@ -2,6 +2,9 @@ package vendingmachine.domain;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
+
+import camp.nextstep.edu.missionutils.Randoms;
 
 public enum Coin {
 	COIN_500(500),
@@ -33,5 +36,16 @@ public enum Coin {
 
 	public static int getAmount(Coin coin) {
 		return coin.amount;
+	}
+
+	public static Map<Coin, Integer> decideCoinRandomly(Map<Coin, Integer> coinMap, Coin[] coins, int balance) {
+		for (int i = 0; i < coins.length - 1; i++) {
+			List<Integer> possibleQuantity = Coin.getPossibleQuantity(coins[i], balance);
+			coinMap.put(coins[i], Randoms.pickNumberInList(possibleQuantity));
+			balance = Coin.calculateResidue(coins[i], balance, coinMap.get(coins[i]));
+		}
+
+		coinMap.put(coins[coins.length - 1], Coin.getMaxQuantity(coins[coins.length - 1], balance));
+		return coinMap;
 	}
 }
